python/pyproject.toml
[tool.poetry]
name = "consensus-python"
version = "0.1.0"
description = "Python package for vector indexing and recommendations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
supabase = "^2.0.0"
psycopg = {extras = ["binary"], version = "^3.1.0"}
pgvector = "^0.2.0"
python-dotenv = "^1.0.0"
numpy = "^1.24.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.0.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Alternative: If you prefer requirements.txt instead of Poetry, uncomment below
# and create a requirements.txt file with:
# supabase>=2.0.0
# psycopg[binary]>=3.1.0
# pgvector>=0.2.0
# python-dotenv>=1.0.0
# numpy>=1.24.0

